Add popup hover in nft marketpace, nft explorer, wallet inventory
