+++ pages/nft-market/nft-marketplace.vue
@@ -1,33 +1,35 @@
-<template lang="pug">
-.j-container
-  div
-    nuxt-link(:to='"/nft-market"', :exact='true')
-      a#return-btn Return
-    h4 Marketplace
-  MarketTab(
-    :data='data',
-    :currentTab='currentTab',
-    :handleTab='handleTab',
-    :handleSearch='handleSearch',
-    :collectionData='collectionData',
-    :handleCollection='handleCollection',
-    :searchValue='search',
-    :handleSearchValue='handleSearchValue'
-  )
-  .grid-container(v-if='loading')
-    CustomSkeletonVue(v-for='item in 10', :key='item', :height="330")
-  .grid-container(v-else)
-    .d-flex.justify-content-center(
-      v-for='(item, index) in marketData',
-      :key='index'
+<template lang='pug'>
+  .j-container
+    div
+      nuxt-link(:to='"/nft-market"', :exact='true')
+        a#return-btn Return
+      h4 Marketplace
+    MarketTab(
+      :data='data',
+      :currentTab='currentTab',
+      :handleTab='handleTab',
+      :handleSearch='handleSearch',
+      :collectionData='collectionData',
+      :handleCollection='handleCollection',
+      :searchValue='search',
+      :handleSearchValue='handleSearchValue'
     )
-      NormalCard(
-        v-if='item',
-        :data='item',
-        :price='getPrice',
-        :kindBut='currentTab',
-        mode='market'
+    .grid-container(v-if='loading')
+      CustomSkeletonVue(v-for='item in 10', :key='item', :height="330")
+    .grid-container(v-else)
+      .d-flex.justify-content-center(
+        v-for='(item, index) in marketData',
+        :key='index'
       )
+        NormalCard(
+          v-if='item',
+          :data='item',
+          :price='getPrice',
+          :kindBut='currentTab',
+          mode='market',
+          :getOverData='getInventoryOverData'
+          :overData='overData'
+        )
 </template>
 
 <script>
@@ -44,7 +46,7 @@ export default {
   components: {
     NormalCard,
     MarketTab,
-    CustomSkeletonVue,
+    CustomSkeletonVue
   },
 
   data() {
@@ -58,18 +60,31 @@ export default {
       data: {
         searchIcon: searchImg,
         filterIcon: filterImg,
-        downIcon: downImg,
+        downIcon: downImg
       },
       limit: 40,
+      overData: ''
     }
   },
   computed: {
-    ...mapState(['network']),
+    ...mapState(['network', 'user']),
     ...mapState('wallet', ['systemPrice']),
     getPrice() {
       let price = this.systemPrice
       return price
-    },
+    }
+  },
+
+  watch: {
+    currentTab(newCurrnetTab, oldCurrentTab) {
+      this.currentCollectionName = ''
+      this.search = ''
+      if (newCurrnetTab === 'sales') {
+        this.getSaleData()
+      } else {
+        this.getAuctionData()
+      }
+    }
   },
 
   mounted() {
@@ -80,6 +95,45 @@ export default {
   methods: {
     handleTab(value) {
       this.currentTab = value
+    },
+    async getInventoryOverData(params) {
+      this.overData = ''
+      const templateStats = await this.getTemplateStats(params.collection_name, params.template_id)
+      const specificAsset = await this.getSpecificAsset(params.collection_name, params.template_id)
+      const assetsSales = await this.getAssetsSales(params.asset_id)
+      const saleData = await this.getOverSale(params.collection_name, params.template_id)
+      this.overData = {
+        templateStats,
+        specificAsset,
+        assetsSales,
+        saleData
+      }
+    },
+
+    async getTemplateStats(collection_name, template_id) {
+      return await this.$store.dispatch('api/getTemplateStats', {
+        collection_name, template_id
+      })
+    },
+
+    async getSpecificAsset(collection_name, template_id) {
+      return await this.$store.dispatch('api/getAssets', {
+        owner: this.user.name,
+        collection_name,
+        template_id
+      })
+    },
+
+    async getAssetsSales(asset_id) {
+      return await this.$store.dispatch('api/getAssetsSales', {
+        asset_id
+      })
+    },
+
+    async getOverSale(collectionName, template_id) {
+      return await this.$store.dispatch('api/getSaleData', {
+        collectionName, template_id, symbol: 'WAX', state: 1, limit: 1
+      })
     },
 
     handleSearchValue(value) {
@@ -105,7 +159,7 @@ export default {
 
     async getCollectionData() {
       const data = await this.$store.dispatch('api/getCollectionData', {
-        author: '',
+        author: ''
       })
       this.collectionData = data
     },
@@ -115,7 +169,7 @@ export default {
       const data = await this.$store.dispatch('api/getSaleData', {
         limit: this.limit,
         search: this.search,
-        collectionName: this.currentCollectionName,
+        collectionName: this.currentCollectionName
       })
       this.marketData = data
       this.loading = false
@@ -125,23 +179,11 @@ export default {
       const data = await this.$store.dispatch('api/getAuctionData', {
         limit: this.limit,
         search: this.search,
-        collectionName: this.currentCollectionName,
+        collectionName: this.currentCollectionName
       })
       this.marketData = data
       this.loading = false
-    },
-  },
-
-  watch: {
-    currentTab(newCurrnetTab, oldCurrentTab) {
-      this.currentCollectionName = ''
-      this.search = ''
-      if (newCurrnetTab === 'sales') {
-        this.getSaleData()
-      } else {
-        this.getAuctionData()
-      }
-    },
+    }
   },
 
   head() {
@@ -152,18 +194,19 @@ export default {
         {
           hid: 'description',
           name: 'description',
-          content: 'Atomic, no fee, NFT marketplace.',
-        },
-      ],
+          content: 'Atomic, no fee, NFT marketplace.'
+        }
+      ]
     }
-  },
+  }
 }
 </script>
 
-<style lang="scss">
+<style lang='scss'>
 #return-btn::before {
   content: '←';
 }
+
 #return-btn {
   font-weight: 500;
   font-size: 14px;
@@ -171,9 +214,11 @@ export default {
   cursor: pointer;
   padding-left: 10px;
 }
+
 h4 {
   margin: 32px 0;
 }
+
 div.grid-container {
   display: grid;
   grid-template-columns: auto auto auto auto;
